requirements.txt
# FinanceGPT — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework (Backend) ──────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.0

# ── Frontend ─────────────────────────────────────────────────
streamlit==1.32.0

# ── LLM — Groq (free, fastest) ───────────────────────────────
groq==0.5.0

# ── Vector Database ───────────────────────────────────────────
qdrant-client==1.8.0

# ── Embeddings (multilingual, runs locally, free) ────────────
sentence-transformers==2.7.0
torch==2.2.1          # needed by sentence-transformers

# ── HTTP Requests (frontend → backend) ───────────────────────
requests==2.31.0

# ── Environment Variables ────────────────────────────────────
python-dotenv==1.0.1

# ── Data Processing ──────────────────────────────────────────
pydantic==2.6.1        # used by FastAPI for data validation
